Add global flag exclude_stopwords (fixes issue #8)
The flag will enable stopwords filtering based on stop-words create.

The list of stopwords is importy with the crate stop-words. It contains
a list of ISO and NLTK stopwords and multiple languages.

Right now the default language is English.

The stopwords gets added to the existing exclude struct as a hashset. It
will by default ignore capitalisation so words like `The` and `the` both
get matched.

Examples:

Without filter:

```
freq <<< "The bird is the word"

0.2 - 1 - The
0.2 - 1 - bird
0.2 - 1 - word
0.2 - 1 - is
0.2 - 1 - the
```

With filter:

```
freq --exclude-stopwords <<< "The bird is the word"

0.5 - 1 - word
0.5 - 1 - bird
```